Spring_core_basic/project/corebasic1/src/test/java/com/kloong/corebasic1/discount/RateDiscountPolicyTest.java
package com.kloong.corebasic1.discount;

import com.kloong.corebasic1.member.Grade;
import com.kloong.corebasic1.member.Member;
import org.junit.jupiter.api.DisplayName;
import org.junit.jupiter.api.Test;

import static org.assertj.core.api.Assertions.*;

class RateDiscountPolicyTest {

    DiscountPolicy discountPolicy = new RateDiscountPolicy();

    @Test
    @DisplayName("VIP는 10% 할인이 적용되어야 한다")
    void vip_discount() {
        //given
        Member member = new Member(1L, "meberVIP", Grade.VIP);

        //when
        int discount = discountPolicy.discount(member, 10000);

        //then
        assertThat(discount).isEqualTo(1000);
    }

    @Test
    @DisplayName("VIP가 아니면 할인이 적용되지 않아야 한다")
    void basic_discount()
    {
        //given
        Member member = new Member(2L, "meberBASIC", Grade.BASIC);

        //when
        int discount = discountPolicy.discount(member, 10000);

        //then
        assertThat(discount).isEqualTo(0);
    }

}